% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_function.R
\name{calc_percentiles}
\alias{calc_percentiles}
\title{Calculate percentiles for drought statistics}
\usage{
calc_percentiles(historical_flow, realtime_data, expected, number_of_years)
}
\arguments{
\item{historical_flow}{historical flow data from hydat database}

\item{realtime_data}{realtime data (from last 30 days)}

\item{expected}{percentile bins}

\item{number_of_years}{number of years of data}
}
\value{
Returns a data frame with percentiles based on historic data
calc_percentiles()
}
\description{
Function for calculating the percentile from tidyhydat extracted data. Limited to current day calculation. Only calculates percentiles if there is at least 5 years of historic data. Made Aug 2020 by Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
