% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_wc_function.R
\name{drought_stats_wc}
\alias{drought_stats_wc}
\title{Function for rendering the West Coast RMarkdown product}
\usage{
drought_stats_wc(basin, save_loc = "/inst/")
}
\arguments{
\item{basin}{The basin you want to render the drought statistics html page for}

\item{save_loc}{The path that you want to save a copy of the html file at. Defaults to '/inst/'}
}
\value{
The function returns a html version of the drought product
}
\description{
Function for rendering the West Coast RMarkdown product
}
\keyword{drought}
