% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapedroughtmap_function.R
\name{scrape_drought_map}
\alias{scrape_drought_map}
\title{Scrape drought map for drought levels and polygons}
\usage{
scrape_drought_map()
}
\value{
Returns the polygons and drought levels from the current BC Drought Portal
scrape_drought_map()
}
\description{
Function for scraping drought polygons and levels from BC Drought Portal
}
\examples{
\dontrun{}
}
\keyword{internal}
