% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{start.time}
\alias{start.time}
\title{Function to get a new token if process takes too long (for tidyhydat.ws())}
\format{
An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.
}
\usage{
start.time
}
\description{
Function to get a new token if process takes too long (for tidyhydat.ws())
}
\examples{
\dontrun{}
}
\keyword{get}
\keyword{new}
\keyword{token}
