% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_temp_function.R
\name{water_temp}
\alias{water_temp}
\title{Function for getting stream and lake temperatures from the WS sites}
\usage{
water_temp(station_wt)
}
\arguments{
\item{station_wt}{Stations that you want to calculate drought-relevant statistics for. Function will also retrieve data itself}
}
\value{
Returns the temperature data for sites retrieved from tidyhydat.ws
}
\description{
Function for getting stream and lake temperatures from the WS sites
}
\keyword{drought}
