% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_temp_function.R
\name{water_temp_aquarius}
\alias{water_temp_aquarius}
\title{Get water temperatures from AQUARIUS}
\usage{
water_temp_aquarius(station_wt, as_drought_stats = TRUE)
}
\arguments{
\item{station_wt}{Stations that you want to calculate drought-relevant statistics for. Function will also retrieve data itself}
}
\value{
Returns the temperature data for sites retrieved from tidyhydat
}
\description{
Function for getting stream and lake temperatures from provincial aquarius sites
}
\examples{
\dontrun{}
}
\keyword{functions}
\keyword{temperature}
\keyword{water}
