% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bc_logo.R
\name{download_bc_logo}
\alias{download_bc_logo}
\title{Function for downloading the correct BC government logo for superimposing on the snow basin index map}
\usage{
download_bc_logo()
}
\description{
Function for downloading the correct BC government logo for superimposing on the snow basin index map
}
\examples{
download_bc_logo()
}
\keyword{BC}
\keyword{Download}
\keyword{governemnt}
\keyword{logo}
