% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droughtstats_function.R
\name{drought_statistics}
\alias{drought_statistics}
\title{calculate all drought relevant statistics for stations}
\usage{
drought_statistics(stations)
}
\arguments{
\item{stations}{A list of stations that you want to calculate drought-relevant statistics for. Function will also retrieve data itself}
}
\value{
The function returns a dataframe of statistics for the list of stations provided
}
\description{
Function for assembling drought statistics and data directly from tidyhydat. Initially developed Aug 2020 by Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{drought}
\keyword{stats}
